[#36342653] Fixing FileDownload Helper for Selenium tests.
The FileDownload helper crapped out because a new exception was added for when json couldn't be parsed correctly. Therefore, I made it such that you must specify that json is what you want otherwise, raw text is sent back. With that said, json is the client's default. You must change the format property in order to change the way it's parsed.